Migrate from Preact to React
As a developer, I want to migrate the entire project from Preact to React so that we can standardize our development on a single framework and access additional functionalities specific to React, which are not available with Preact. This will ensure our application can integrate future features that leverage the full capabilities of React.

**Tasks**

- [x] Remove Preact and install React along with ReactDOM in the project dependencies.
- [x] Update import statements in view and modal classes from Preact to React equivalents.
- [x] Refactor the `render()` method implementations to align with React standards.
- [x] Adjust the project's build and configuration files to support React.
- [x] Conduct thorough testing to ensure all existing features function as intended with React.